+++ apps/backend/src/routers/oauth/userinfo.ts
@@ -1,0 +1,68 @@
+import express from "express";
+
+import { oauthRepository } from "../../db/repositories";
+
+const userinfoRouter = express.Router();
+
+/**
+ * OAuth 2.0 UserInfo Endpoint
+ * Returns information about the authenticated user
+ */
+userinfoRouter.get("/oauth/userinfo", async (req, res) => {
+  try {
+    const authHeader = req.headers.authorization;
+
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return res.status(401).json({
+        error: "invalid_token",
+        error_description: "Missing or invalid authorization header",
+      });
+    }
+
+    const token = authHeader.substring(7); // Remove "Bearer " prefix
+
+    // Validate MCP token format
+    if (!token.startsWith("mcp_token_")) {
+      return res.status(401).json({
+        error: "invalid_token",
+        error_description: "Invalid access token format",
+      });
+    }
+
+    // Look up token data (in production, this should validate signature and lookup in database)
+    const tokenData = await oauthRepository.getAccessToken(token);
+    if (!tokenData) {
+      return res.status(401).json({
+        error: "invalid_token",
+        error_description: "Token not found or expired",
+      });
+    }
+
+    // Check if token has expired
+    if (Date.now() > tokenData.expires_at.getTime()) {
+      await oauthRepository.deleteAccessToken(token);
+      return res.status(401).json({
+        error: "invalid_token",
+        error_description: "Access token has expired",
+      });
+    }
+
+    // For MCP tokens, return basic user info based on the user_id stored with the token
+    // In a real implementation, you would fetch actual user data from the database
+    res.json({
+      sub: tokenData.user_id,
+      email: `user-${tokenData.user_id}@metamcp.local`,
+      name: `MetaMCP User ${tokenData.user_id}`,
+      preferred_username: `user_${tokenData.user_id}`,
+      scope: tokenData.scope,
+    });
+  } catch (error) {
+    console.error("Error in OAuth userinfo endpoint:", error);
+    res.status(500).json({
+      error: "server_error",
+      error_description: "Internal server error",
+    });
+  }
+});
+
+export default userinfoRouter;
