+++ apps/backend/src/routers/oauth/utils.ts
@@ -1,0 +1,78 @@
+import express from "express";
+
+// OAuth 2.0 Authorization Parameters interface
+export interface OAuthParams {
+  client_id: string;
+  redirect_uri: string;
+  scope?: string;
+  state?: string;
+  code_challenge?: string;
+  code_challenge_method?: string;
+}
+
+/**
+ * Helper function to get the correct base URL from request
+ * Prioritizes APP_URL environment variable, then checks proxy headers
+ */
+export function getBaseUrl(req: express.Request): string {
+  // Prioritize APP_URL environment variable
+  if (process.env.APP_URL) {
+    return process.env.APP_URL;
+  }
+
+  // Check for forwarded headers from Next.js proxy
+  const forwardedHost = req.headers["x-forwarded-host"] as string;
+  const forwardedProto = req.headers["x-forwarded-proto"] as string;
+
+  if (forwardedHost) {
+    const protocol = forwardedProto || "http";
+    return `${protocol}://${forwardedHost}`;
+  }
+
+  // Fallback to request host
+  return `${req.protocol}://${req.get("host")}`;
+}
+
+/**
+ * Middleware to add JSON parsing for OAuth POST endpoints
+ */
+export function jsonParsingMiddleware(
+  req: express.Request,
+  res: express.Response,
+  next: express.NextFunction,
+) {
+  // Only apply JSON parsing for OAuth POST endpoints that need parsed body
+  const needsJsonParsing =
+    (req.path.startsWith("/oauth/") && req.method === "POST") ||
+    (req.path === "/oauth/register" && req.method === "POST");
+
+  if (needsJsonParsing) {
+    return express.json({
+      limit: "10mb",
+      type: "application/json",
+    })(req, res, next);
+  }
+  next();
+}
+
+/**
+ * Middleware to add URL-encoded form parsing for OAuth POST endpoints
+ */
+export function urlencodedParsingMiddleware(
+  req: express.Request,
+  res: express.Response,
+  next: express.NextFunction,
+) {
+  // Only apply URL-encoded parsing for OAuth POST endpoints
+  const needsUrlencodedParsing =
+    (req.path.startsWith("/oauth/") && req.method === "POST") ||
+    (req.path === "/oauth/register" && req.method === "POST");
+
+  if (needsUrlencodedParsing) {
+    return express.urlencoded({
+      extended: true,
+      limit: "10mb",
+    })(req, res, next);
+  }
+  next();
+}
