+++ apps/backend/src/routers/public-metamcp.ts
@@ -2,7 +2,7 @@ import cors from "cors";
 import express from "express";
 
 import { endpointsRepository } from "../db/repositories/endpoints.repo";
-import oauthMetadataRouter from "./public-metamcp/oauth";
+import oauthRouter from "./oauth";
 import openApiRouter from "./public-metamcp/openapi";
 import sseRouter from "./public-metamcp/sse";
 import streamableHttpRouter from "./public-metamcp/streamable-http";
@@ -34,7 +34,7 @@ publicEndpointsRouter.use((req, res, next) => {
 });
 
 // OAuth metadata discovery endpoints (must be mounted first for .well-known paths)
-publicEndpointsRouter.use(oauthMetadataRouter);
+publicEndpointsRouter.use(oauthRouter);
 
 // Use StreamableHTTP router for /mcp routes
 publicEndpointsRouter.use(streamableHttpRouter);
